refactor(mcp): extract calculator engine
Move deterministic calculator implementation behind a dedicated translation unit while preserving the tool wrapper, compatibility behavior, and safe-headless source manifests.
