+++ src/standalone/src/core/mcp/calculator_engine.hpp
@@ -1,0 +1,8 @@
+#pragma once
+
+#include "calculator_tool.hpp"
+
+namespace mcp_standalone::ida_compat
+{
+    tool_result_t calculate_engine(const json& params, const workspace_request_context_t& context);
+}
